+++ tests/test_sap/test_sap_login.py
@@ -4,7 +4,11 @@ import os
 import unittest
 from tkinter import simpledialog
 
+from dotenv import load_dotenv, set_key
+
 from itk_dev_shared_components.sap import sap_login
+
+load_dotenv()
 
 
 class TestSapLogin(unittest.TestCase):
@@ -15,7 +19,7 @@ class TestSapLogin(unittest.TestCase):
         """Show popups that asks for username, password and new password
         used in the following tests.
         """
-        cls.username, cls.password = os.environ['SAP Login'].split(';')
+        cls.username, cls.password = os.environ['SAP_LOGIN'].split(';')
         cls.new_password = simpledialog.askstring("Enter new password", "Enter the new password to be used in testing the change SAP password function.\nRemember to write down the new password!\nLeave empty to skip test_change_password.")
 
     def setUp(self) -> None:
@@ -38,7 +42,7 @@ class TestSapLogin(unittest.TestCase):
     def test_change_password(self):
         """Test the function change password.
         Due to a limit in SAP you can only run this function once per day.
-        If no new_password is entered in the setup
+        If no new_password is entered in the setup, test is skipped.
         """
         if not self.new_password:
             raise unittest.SkipTest("Test not run because new_password was missing.")
@@ -46,7 +50,8 @@ class TestSapLogin(unittest.TestCase):
         sap_login.change_password(self.username, self.password, self.new_password)
         # Change password for all coming tests
         self.password = self.new_password
-        os.environ['SAP Login'] = f"{self.username};{self.password}"
+        os.environ['SAP_LOGIN'] = f"{self.username};{self.password}"
+        set_key(".env", 'SAP_LOGIN', self.password)
 
         with self.assertRaises(ValueError):
             sap_login.change_password(self.username, "Foo", self.new_password)